requirements.txt
aiogram==3.4.1
telethon==1.34.0
openai==1.12.0
sqlalchemy[asyncio]==2.0.27
aiosqlite==0.19.0
python-dotenv==1.0.1
pydantic-settings==2.2.1